vo_vaapi: remove loop when bob deinterlacing is used
Bob deinterlacing should normally double the framerate (because it
displays each field separately), but vo_vaapi is pretty much too dumb
to communicate this to the playloop. Instead, it shows both fields at
the same time, which doesn't seem to make any sense. The original
mplayer-vaapi patch did it this way. It turns out showing them both
at the same time can actually produce flicker, so remove this complete
non-sense garbage.

the VO interface can actually handle it, and vo_vdpau is using it. But
since vf_vavpp handles this correctly as well, and there's quite some
management boilerplate involved when doing this in the VO, we just
ignore this and show only one field.