talloc: change talloc destructor signature
Change talloc destructor so that they never fail, and don't return a
status code. This makes our talloc copy even more incompatible to
upstream talloc, but on the other hand this is preparation for getting
rid of talloc entirely. (The talloc replacement in the next commit won't
the talloc_free equivalent allow to fail, and the destructor return
value would be useless. But I don't want to change any mpv code either;
the idea is that the talloc replacement commit can be reverted for some
time in order to test whether the talloc replacement introduced a
regression.)