Copyright, LICENSE: switch to GPL version 2 or later
Now that talloc has been removed, the license can be switched back to
GPLv2+. Actually, there never was a GPLv2+ licensed MPlayer (fork or
not) until now, but since some code has been removed, this is possible
now. Rewrite the Copyright file to explain the reasons for the licenses
MPlayer and forks use. The Copyright file didn't contain anything
interesting anymore, and all information it contained is available at
other places in the source tree.

The reason for the license change is that it should improve
interoperability with differently licensed code in general.

This essentially reverts commit 1752808.